// web/src/api/generated/models/ChapterContentItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ChapterContentItem
 */
export interface ChapterContentItem {
    /**
     * 
     * @type {string}
     * @memberof ChapterContentItem
     */
    targetId: string;
    /**
     * 
     * @type {ChapterContentItemTargetTypeEnum}
     * @memberof ChapterContentItem
     */
    targetType: ChapterContentItemTargetTypeEnum;
}


/**
 * @export
 */
export const ChapterContentItemTargetTypeEnum = {
    MEMORY: 'MEMORY',
    HEART_MOMENT: 'HEART_MOMENT',
    MILESTONE: 'MILESTONE'
} as const;
export type ChapterContentItemTargetTypeEnum = typeof ChapterContentItemTargetTypeEnum[keyof typeof ChapterContentItemTargetTypeEnum];


/**
 * Check if a given object implements the ChapterContentItem interface.
 */
export function instanceOfChapterContentItem(value: object): value is ChapterContentItem {
    if (!('targetId' in value) || value['targetId'] === undefined) return false;
    if (!('targetType' in value) || value['targetType'] === undefined) return false;
    return true;
}

export function ChapterContentItemFromJSON(json: any): ChapterContentItem {
    return ChapterContentItemFromJSONTyped(json, false);
}

export function ChapterContentItemFromJSONTyped(json: any, ignoreDiscriminator: boolean): ChapterContentItem {
    if (json == null) {
        return json;
    }
    return {
        
        'targetId': json['targetId'],
        'targetType': json['targetType'],
    };
}

export function ChapterContentItemToJSON(json: any): ChapterContentItem {
    return ChapterContentItemToJSONTyped(json, false);
}

export function ChapterContentItemToJSONTyped(value?: ChapterContentItem | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'targetId': value['targetId'],
        'targetType': value['targetType'],
    };
}